Converted the lookuptable to use floats instead of unsigned ints ...

+++ lookuptable/lut.h
@@ -1,1 +1,1 @@
-void lookuptable(int *, unsigned int, unsigned int, unsigned int *, unsigned int *, unsigned int);
+void lookuptable(int *, unsigned int, unsigned int, float *, float *, unsigned int);
